+++ libraries/provider_spotify_app.rb
@@ -3,7 +3,7 @@
 # Cookbook Name:: spotify
 # Library:: provider_spotify_app
 #
-# Copyright 2015 Jonathan Hartman
+# Copyright 2015-2016, Jonathan Hartman
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -67,8 +67,8 @@ class Chef
       # @raise [NotImplementedError] if not defined for this provider.
       #
       def install!
-        fail(NotImplementedError,
-             "`install!` method not implemented for #{self.class} provider")
+        raise(NotImplementedError,
+              "`install!` method not implemented for #{self.class} provider")
       end
 
       #
@@ -77,8 +77,8 @@ class Chef
       # @raise [NotImplementedError] if not defined for this provider.
       #
       def remove!
-        fail(NotImplementedError,
-             "`remove!` method not implemented for #{self.class} provider")
+        raise(NotImplementedError,
+              "`remove!` method not implemented for #{self.class} provider")
       end
     end
   end