+++ libraries/provider_spotify_app_debian.rb
@@ -3,7 +3,7 @@
 # Cookbook Name:: spotify
 # Library:: provider_spotify_app_debian
 #
-# Copyright 2015 Jonathan Hartman
+# Copyright 2015-2016, Jonathan Hartman
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -46,9 +46,8 @@ class Chef
         #
         def install!
           add_repo
-          package 'spotify-client' do
-            action :install
-          end
+          package 'xdg-utils'
+          package 'spotify-client'
         end
 
         #