+++ metadata.rb
@@ -1,6 +1,5 @@
 # Encoding: UTF-8
-#
-# rubocop:disable SingleSpaceBeforeFirstArg
+
 name             'spotify'
 maintainer       'Jonathan Hartman'
 maintainer_email 'j@p4nt5.com'
@@ -11,10 +10,9 @@ version          '0.1.1'
 
 depends          'dmg', '~> 2.2'
 depends          'windows', '~> 1.37'
-depends          'apt', '~> 2.7'
+depends          'apt', '~> 3.0'
 
 supports         'mac_os_x'
 supports         'windows'
 supports         'ubuntu'
 supports         'debian'
-# rubocop:enable SingleSpaceBeforeFirstArg